--- Grafische interface 2.0/socket_gegevens.json
@@ -1,1 +1,0 @@
-{"client_socket":"<socket.socket fd=504, family=AddressFamily.AF_INET, type=SocketKind.SOCK_STREAM, proto=0, laddr=('192.168.137.1', 53930), raddr=('192.168.137.110', 5555)>","ip_address":"192.168.137.110","port":"5555","uuid":"test","is_gui":"False","connected_to_server":"True","registered":"False","debug":"True","last_ping":"1516708552.1170208","barrier_open":"False","connection_handler":"<socket.socket fd=504, family=AddressFamily.AF_INET, type=SocketKind.SOCK_STREAM, proto=0, laddr=('192.168.137.1', 53930), raddr=('192.168.137.110', 5555)>","online":"False"}
